+++ celery_worker.py
@@ -1,0 +1,4 @@
+from webapp import celery, create_app
+
+app = create_app()
+app.app_context().push()
