+++ webapp/celery_app.py
@@ -1,0 +1,6 @@
+from celery import Celery
+
+from webapp.config import CELERY_BROKER_URL, CELERY_RESULT_BACKEND
+
+celery = Celery(__name__, backend=CELERY_RESULT_BACKEND, broker=CELERY_BROKER_URL)
+celery.autodiscover_tasks()
