+++ webapp/config.py
@@ -23,3 +23,5 @@ YOOMONEY_WALLET = os.environ['YOOMONEY_WALLET']
 YOOMONEY_SECRET_KEY = os.environ['YOOMONEY_SECRET_KEY']
 MAINSMS_PROJECT_NAME = os.environ['MAINSMS_PROJECT_NAME']
 MAINSMS_API_KEY = os.environ['MAINSMS_API_KEY']
+CELERY_BROKER_URL = "redis://localhost:6379"
+CELERY_RESULT_BACKEND = "redis://localhost:6379"
