+++ webapp/templates/menu.html
@@ -47,7 +47,7 @@
                             <div class="block-wishlist action">
                                 <a class="icon-link" href="{{ url_for('marketplace.favorite_page') }}">
                                 <i class="flaticon-heart"></i>
-                                <span class="count">{{ count_favorite_products_current_user() }}</span>
+                                <span class="count">{{ count_favorite_products_current_user }}</span>
                                 <span class="text">
                                 <span class="sub">Избранное</span>
                                 Мой список желаний</span>
